refactor(test): extract `RecordingServerTestUtil` for reusable test utilities
