+++ core/src/test/java/com/predic8/membrane/core/interceptor/soap/SoapOperationExtractorTest.java
@@ -23,8 +23,8 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import static com.predic8.membrane.core.http.Request.post;
 import static com.predic8.membrane.core.interceptor.soap.SoapOperationExtractor.SOAP_OPERATION;
 import static com.predic8.membrane.core.interceptor.soap.SoapOperationExtractor.SOAP_OPERATION_NS;
-import static com.predic8.membrane.core.util.xml.parser.HardenedSaxParserTest.freePort;
-import static com.predic8.membrane.core.util.xml.parser.HardenedSaxParserTest.startRecordingServer;
+import static com.predic8.membrane.core.util.RecordingServerTestUtil.freePort;
+import static com.predic8.membrane.core.util.RecordingServerTestUtil.startRecordingServer;
 import static org.junit.jupiter.api.Assertions.*;
 
 public class SoapOperationExtractorTest {
